Change NodeKey to i32
To avoid accidental mixups with other usages of usize.

+++ src/main.rs
@@ -45,7 +45,7 @@ struct Selection {
     pos: TextPos,
 }
 
-type NodeKey = usize;
+type NodeKey = i32;
 
 enum Line {
     Text {
